+++ commands/addmoney.js
@@ -6,7 +6,7 @@ module.exports.run = async (client, message) => {
         if (message.author.id !== globalVars.ownerID) {
             return message.reply(globalVars.lackPerms)
         };
-        const { bank } = require('../bank');
+        const { bank } = require('../database/bank');
         const input = message.content.slice(1).trim();
         const [, , commandArgs] = input.match(/(\w+)\s*([\s\S]*)/);
         const currentAmount = bank.currency.getBalance(message.author.id);
