+++ commands/shop.js
@@ -1,5 +1,5 @@
 exports.run = async (client, message) => {
-    const {CurrencyShop} = require('../storeObjects');
+    const {CurrencyShop} = require('../database/dbObjects');
     const items = await CurrencyShop.findAll();
     return message.channel.send(items.map(i => `${i.name}: ${i.cost}💰`).join('\n'), { code: true });
 };
