+++ commands/switch.js
@@ -3,7 +3,7 @@ module.exports.run = async (client, message) => {
         // Import globals
         let globalVars = require('../events/ready');
 
-        const { bank } = require('../bank');
+        const { bank } = require('../database/bank');
         const input = message.content.slice(1).trim();
         const [, , arguments] = input.match(/(\w+)\s*([\s\S]*)/);
         if (arguments.length<1) return message.channel.send(`> Please specify a valid Nintendo Switch friend code, ${message.author}.`);
