+++ commands/transfer.js
@@ -1,6 +1,6 @@
 exports.run = async (client, message) => {
     try {
-        const { bank } = require('../bank');
+        const { bank } = require('../database/bank');
         const input = message.content.slice(1).trim();
         const [, , commandArgs] = input.match(/(\w+)\s*([\s\S]*)/);
         const currentAmount = bank.currency.getBalance(message.author.id);
