+++ events/ready.js
@@ -1,8 +1,8 @@
 module.exports = async (client) => {
   try {
     const stan = require('../stan')(client);
-    const { bank } = require('../bank');
-    const { Users } = require('../storeObjects');
+    const { bank } = require('../database/bank');
+    const { Users } = require('../database/dbObjects');
     const storedBalances = await Users.findAll();
     storedBalances.forEach(b => bank.currency.set(b.user_id, b));
     console.log(`Ready to serve in ${client.channels.cache.size} channels on ${client.guilds.cache.size} servers, for a total of ${client.users.cache.size} users.`);
