--- storeObjects.js
@@ -1,36 +1,0 @@
-const Sequelize = require('sequelize');
-
-const sequelize = new Sequelize('database', 'username', 'password', {
-	host: 'localhost',
-	dialect: 'sqlite',
-	logging: false,
-	storage: 'database.sqlite',
-});
-
-const Users = require('./models/Users')(sequelize, Sequelize.DataTypes);
-const CurrencyShop = require('./models/CurrencyShop')(sequelize, Sequelize.DataTypes);
-const UserItems = require('./models/UserItems')(sequelize, Sequelize.DataTypes);
-const EligibleRoles = require('./models/EligibleRoles')(sequelize, Sequelize.DataTypes);
-UserItems.belongsTo(CurrencyShop, { foreignKey: 'item_id', as: 'item' });
-
-Users.prototype.addItem = async function(item) {
-	const useritem = await UserItems.findOne({
-		where: { user_id: this.user_id, item_id: item.id },
-	});
-
-	if (useritem) {
-		useritem.amount += 1;
-		return useritem.save();
-	}
-
-	return UserItems.create({ user_id: this.user_id, item_id: item.id, amount: 1 });
-};
-
-Users.prototype.getItems = function() {
-	return UserItems.findAll({
-		where: { user_id: this.user_id },
-		include: ['item'],
-	});
-};
-
-module.exports = { Users, CurrencyShop, UserItems, EligibleRoles};
